Add the spi/lcd display code
Fyi I used hardware debouncing for the button (10k resistor and 100 nF cap), may look into software debouncing later. Not sure how well this will play with the led matrix stuff. Basically this should have the display scroll through "select difficulty level" until the button is pressed, at which point you will be able to loop through the levels 1-5 (stored in the int variable "difficulty"). Ideally this should probably reset after a pull of the lever, and maybe we can have a message if you win or lose or something.